:star: add get game map from htpp client

+++ Server/Player.cpp
@@ -16,7 +16,7 @@ void Player::setName(std::string name)
 
 int Player::sendMessage(std::string mess)
 {
-	int iSendResult = send(_clientSocket, mess.c_str(), mess.size(), 0);
+	int iSendResult = send(_clientSocket, mess.c_str(), (int)mess.size(), 0);
 	if (iSendResult == SOCKET_ERROR) {
 		closesocket(_clientSocket);
 		WSACleanup();
